Move contracts to test and remove deploy script.
