+++ packages/solidity-mapper/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "solidity-mapper",
+  "name": "@vulcanize/solidity-mapper",
   "version": "0.1.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -18,7 +18,6 @@
     "typescript": "^4.3.2"
   },
   "scripts": {
-    "deploy": "hardhat run scripts/deploy.ts",
     "build": "tsc",
     "test": "hardhat test"
   }
