+++ packages/solidity-mapper/src/index.ts
@@ -1,1 +1,1 @@
-export { getStorageValue, StorageLayout, GetStorageAt } from './lib/storage';
+export { getStorageValue, StorageLayout, GetStorageAt } from './storage';
