+++ packages/solidity-mapper/test/contracts/TestIntegers.sol
