+++ packages/solidity-mapper/test/contracts/TestStrings.sol
