+++ packages/solidity-mapper/test/utils.ts
@@ -1,8 +1,10 @@
 import { artifacts, ethers } from 'hardhat'
 import { CompilerOutput, CompilerOutputBytecode } from 'hardhat/types';
 
-import { StorageLayout, GetStorageAt } from '../../src';
+import { StorageLayout, GetStorageAt } from '../src';
 
+// storageLayout doesnt exist in type CompilerOutput doesnt.
+// Extending CompilerOutput type to include storageLayout property.
 interface StorageCompilerOutput extends CompilerOutput {
   contracts: {
     [sourceName: string]: {
@@ -29,18 +31,18 @@ export const getStorageLayout = async (contractName: string) => {
   const artifact = await artifacts.readArtifact(contractName);
   const buildInfo = await artifacts.getBuildInfo(`${artifact.sourceName}:${artifact.contractName}`)
 
-  if (buildInfo) {
-    const output: StorageCompilerOutput = buildInfo.output
-    const { storageLayout } = output.contracts[artifact.sourceName][artifact.contractName];
-
-    if (storageLayout) {
-      return storageLayout
-    }
-
+  if (!buildInfo) {
     throw new Error('storageLayout not present in compiler output.');
   }
 
-  throw new Error(`Contract hasn't been compiled.`);
+  const output: StorageCompilerOutput = buildInfo.output
+  const { storageLayout } = output.contracts[artifact.sourceName][artifact.contractName];
+
+  if (!storageLayout) {
+    throw new Error(`Contract hasn't been compiled.`);
+  }
+
+  return storageLayout;
 }
 
 /**
@@ -50,5 +52,6 @@ export const getStorageLayout = async (contractName: string) => {
  */
 export const getStorageAt: GetStorageAt = async (address, position) => {
   const value = await ethers.provider.getStorageAt(address, position);
+
   return value;
 }
